Tasj

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 /**
  * main - Entry point
- * Return 0
+ * Return: 0
  */
 
 int main(void)
 {
-	int sum = 0;
+	long int sum = 0;
 	int first = 0, second = 1, i;
 
 	printf("%d, ", first);
 	for (i = 0; i <= 50; i++)
 	{
 		sum = first + second;
-		printf("%d, ", sum);
+		printf("%li, ", sum);
 		first = second;
 		second = sum;
